requirements.txt
# Data manipulation
pandas==2.1.4
numpy==1.26.2
pyarrow==14.0.1

# Snowflake - with pandas support
snowflake-connector-python[pandas]==3.6.0
snowflake-sqlalchemy==1.5.1

# dbt
dbt-core==1.7.4
dbt-snowflake==1.7.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
cryptography==41.0.7

# Development
black==23.12.1
flake8==7.0.0
pytest==7.4.3
